Work on integrating the logger with the new login systems
Created a new method to add to the log that prints a timestamp, the calling class name, and the information.
Moved the uploading of the log into the logger itself.
Made the lab manager simply defer to the login manager, instead of starting the log upload itself, allowing for multiple labs to be done before uploading a log.
Add logger calls throughout  LoginManager, Authenticate, DownloadUtility, LabManager, and TestWrite.