Web Loading with the Bridge is Running
Added some extra logs to the bridge to help determine where content is placed and how big it is
Added more logs to demoSequence, lightingControl, and demo to track application flow
DownloadUtility Now checks if a file already exists before downloading/extractin, and deletes it.
this avoids an error when extracting a file that already exists
Changed what the Login Manager sets the downloaded resource zip and folder name to when the download is
from a debug endpoint. May need to change it for production endpoints in the future as well.