Various changes
Work on #11
Removed Bridge spawning primitives instead of using prefabs from the resources folder.
This will likely break some Bridge example scenes, but the main way we use the bridge
never used this feature.
Added ability to spawn objects marked as transmission normally if Transmission is not
set up.
Added some UI that the labmanager triggers at the start of the transmission lab.
It tracks how many peers you are connected to, and the host has the ability to start the lab.
The lab start sequence is not yet syncronized.