Demo.cs Changes
Work on #16
Added flags to the base Activity module object representing if
the current activity is Transmission based, and if the user is the host.
Demo.cs now conditionally creates objects if it is the host
Added code that makes the sunlight utility use RPC calls if the activity is transmission.