Transmission Synced Audio Playback
Added method to demo.cs that handles playing an audio clip, and can be called by Transmission.
Wrapped audio clip playback in demo.cs and demosequence.cs in if statements
that send an RPC message to play audio if transmitting, and are the host.
demosequence.cs uses playback on demo.cs, not itself.